refactor(schemas): migrate Phase 19 endpoint schemas — 4 files, 16 classes (#6042)
- integration_communication.py: migrate TestConnectionRequest, SendMessageRequest,
  ProviderInfo→CommProviderInfo, WebhookMessageRequest → schemas_workflows.py
- knowledge_collaboration.py: migrate KnowledgeScopeFilter, ShareKnowledgeRequest,
  UpdatePermissionsRequest, KnowledgeAccessResponse → schemas_knowledge.py
  (imports VisibilityLevel from knowledge.ownership)
- files.py: migrate FileInfo, DirectoryListing, FileUploadResponse→FilesAPIUploadResponse,
  FileOperation → schemas_system.py (preserves field_validator path traversal check)
- data_storage.py: migrate StorageCategory, StorageStats, CleanupRequest→
  StorageCleanupRequest, CleanupResult → schemas_system.py

Renames to avoid generic-name collisions:
- ProviderInfo → CommProviderInfo (workflows already has ProviderInfo for project_management)
- FileUploadResponse → FilesAPIUploadResponse (knowledge has FileUploadResponse)
- CleanupRequest → StorageCleanupRequest (knowledge has CleanupRequest, audit has AuditCleanupRequest)

schemas_knowledge.py: imports VisibilityLevel from knowledge.ownership
schemas_system.py: gains field_validator import for FileOperation.validate_path

Cleanup: removed unused BaseModel/field_validator/List/DataResponse imports in files.py
and unused KnowledgeAccessResponse/KnowledgeScopeFilter imports in knowledge_collaboration.py
(kept in schemas for completeness but not used at endpoint).
