+++ autobot-backend/api/schemas_knowledge.py
@@ -13,6 +13,7 @@ from typing import Any, Dict, List, Optional, Union
 from pydantic import BaseModel, Field, field_validator
 
 from constants.threshold_constants import CategoryDefaults, QueryDefaults
+from knowledge.ownership import VisibilityLevel
 from type_defs.common import Metadata
 from utils.path_validation import contains_path_traversal
 
@@ -4382,3 +4383,45 @@ class TrainResponse(BaseModel):
     schema_length: Optional[int] = None
     table_count: Optional[int] = None
     error: Optional[str] = None
+
+
+# ---------------------------------------------------------------------------
+# knowledge_collaboration.py schemas
+# ---------------------------------------------------------------------------
+
+
+class KnowledgeScopeFilter(BaseModel):
+    """Filter for knowledge by scope."""
+
+    scope: Optional[VisibilityLevel] = Field(default=None, description="Visibility level to filter by")
+    organization_id: Optional[str] = Field(default=None, description="Organization ID filter")
+    group_ids: Optional[List[str]] = Field(default=None, description="Group IDs to filter by")
+
+
+class ShareKnowledgeRequest(BaseModel):
+    """Request to share knowledge with users or groups."""
+
+    user_ids: Optional[List[str]] = Field(default=None, description="User IDs to share with")
+    group_ids: Optional[List[str]] = Field(default=None, description="Group IDs to share with")
+
+
+class UpdatePermissionsRequest(BaseModel):
+    """Request to update knowledge permissions."""
+
+    visibility: VisibilityLevel = Field(description="New visibility level")
+    organization_id: Optional[str] = Field(default=None, description="Organization ID for org-level knowledge")
+    group_ids: Optional[List[str]] = Field(default=None, description="Group IDs for group-level knowledge")
+
+
+class KnowledgeAccessResponse(BaseModel):
+    """Response with knowledge access details."""
+
+    fact_id: str
+    owner_id: str
+    visibility: VisibilityLevel
+    organization_id: Optional[str] = None
+    group_ids: List[str] = []
+    shared_with: List[str] = []
+    can_edit: bool
+    can_share: bool
+    can_delete: bool
